ComponentMadness/requirements.txt
jsonfield==2.0.2
django_extensions==1.9.9
sorl-thumbnail==12.4.1
pillow==5.0.0
psycopg2==2.7.3.2
requests==2.18.4